requirements.txt
aenum==3.1.15
aiofiles==24.1.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
attrs==24.2.0
azure-cognitiveservices-speech==1.40.0
bidict==0.23.1
blinker==1.9.0
cartesia==1.0.14
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
coloredlogs==15.0.1
dataclasses-json==0.6.7
deepgram-sdk==3.7.6
deprecation==2.1.0
distro==1.9.0
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.5
fastapi-cli==0.0.5
Flask==3.1.0
flask-sock==0.7.0
Flask-SocketIO==5.4.1
flatbuffers==24.3.25
frozenlist==1.5.0
future==1.0.0
gevent==24.11.1
gevent-websocket==0.10.1
greenlet==3.1.1
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.6
httptools==0.6.4
httpx==0.27.2
humanfriendly==10.0
idna==3.10
iniconfig==2.0.0
iterators==0.2.0
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.7.1
linkedin_scraper==2.11.4
llvmlite==0.43.0
loguru==0.7.2
lxml==5.3.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.23.1
mdurl==0.1.2
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
numba==0.60.0
numpy==1.26.4
onnxruntime==1.19.2
openai==1.50.2
outcome==1.3.0.post0
packaging==24.2
pillow==10.4.0
pipecat-ai==0.0.48
pluggy==1.5.0
propcache==0.2.0
protobuf==4.25.5
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
pyloudnorm==0.1.1
PySocks==1.7.1
pytest==8.3.3
pytest-asyncio==0.24.0
python-deepcompare==1.0.1
python-dotenv==1.0.1
python-engineio==4.10.1
python-multipart==0.0.17
python-socketio==5.11.4
PyYAML==6.0.2
requests==2.32.3
resampy==0.4.3
rich==13.9.4
scipy==1.14.1
selenium==4.26.1
shellingham==1.5.4
simple-websocket==1.1.0
sniffio==1.3.1
sortedcontainers==2.4.0
starlette==0.41.2
sympy==1.13.3
tqdm==4.67.0
trio==0.27.0
trio-websocket==0.11.1
typer==0.13.0
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.0
uvloop==0.21.0
watchfiles==0.24.0
websocket-client==1.8.0
websockets==13.1
Werkzeug==3.1.3
wsproto==1.2.0
yarl==1.17.1
zope.event==5.0
zope.interface==7.1.1
